double --> float

+++ src/main_nep_adam/adam.cuh
@@ -23,8 +23,8 @@ public:
   Adam(Parameters&);
 
   // void zero_gradients();
-  void update(double lr, double* gradients); // Update moments and parameters
-  double* get_parameters(); // get current parameters pointer
+  void update(float lr, float* gradients); // Update moments and parameters
+  float* get_parameters(); // get current parameters pointer
   void output_parameters(Parameters& para); // Output parameters
   void initialize_parameters(Parameters& para); // Initialize optimizer parameters
 
@@ -33,11 +33,11 @@ protected:
   std::mt19937 rng;
 
   // ADAM optimizer hyperparameters
-  // double lr = 1e-3f;     // learning rate α
-  double beta1 = 0.9;   // first order moment decay rate β₁
-  double beta2 = 0.999; // second order moment decay rate β₂ 
-  double eps = 1e-8;    // small constant ε to avoid division by zero 
-  double weight_decay = 0.0; // L2 regularization weight
+  // float lr = 1e-3f;     // learning rate α
+  float beta1 = 0.9f;   // first order moment decay rate β₁
+  float beta2 = 0.999f; // second order moment decay rate β₂ 
+  float eps = 1e-8f;    // small constant ε to avoid division by zero 
+  float weight_decay = 0.0f; // L2 regularization weight
 
   // training parameters
   int step = 0; // initial step
@@ -45,12 +45,12 @@ protected:
 
   // CPU vectors
   GPU_Vector<curandState> curand_states; 
-  std::vector<double> parameters;  // Parameters to be optimized θ
-  std::vector<double> m;         // First moment m
-  std::vector<double> v;         // Second moment v (adaptive learning rate)
+  std::vector<float> parameters;  // Parameters to be optimized θ
+  std::vector<float> m;         // First moment m
+  std::vector<float> v;         // Second moment v (adaptive learning rate)
   // GPU vectors
-  GPU_Vector<double> gpu_parameters; // Parameters to be optimized θ
-  GPU_Vector<double> gpu_m; // First moment m
-  GPU_Vector<double> gpu_v; // Second moment v
+  GPU_Vector<float> gpu_parameters; // Parameters to be optimized θ
+  GPU_Vector<float> gpu_m; // First moment m
+  GPU_Vector<float> gpu_v; // Second moment v
 
 };
