+++ src/main_nep_adam/gradients.cuh
@@ -4,9 +4,9 @@
 static void print_memory_info(const char* stage) {
     size_t free_memory, total_memory;
     cudaMemGetInfo(&free_memory, &total_memory);
-    double free_gb = free_memory / (1024.0 * 1024.0 * 1024.0);
-    double total_gb = total_memory / (1024.0 * 1024.0 * 1024.0);
-    double used_gb = total_gb - free_gb;
+    float free_gb = free_memory / (1024.0 * 1024.0 * 1024.0);
+    float total_gb = total_memory / (1024.0 * 1024.0 * 1024.0);
+    float used_gb = total_gb - free_gb;
     
     printf("%s:\n", stage);
     printf("  Free memory:  %.2f GB\n", free_gb);
@@ -23,7 +23,7 @@ struct Gradients {
     Fp_wb.resize(N * number_of_variables_ann * dim);
     // print_memory_info("after Fp_wb.resize");
   }
-  GPU_Vector<double> E_wb_grad;      // energy w.r.t. w0, b0, w1, b1
-  GPU_Vector<double> grad_sum;
-  GPU_Vector<double> Fp_wb;          // gradient of descriptors w.r.t. w0, b0, w1, b1
+  GPU_Vector<float> E_wb_grad;      // energy w.r.t. w0, b0, w1, b1
+  GPU_Vector<float> grad_sum;
+  GPU_Vector<float> Fp_wb;          // gradient of descriptors w.r.t. w0, b0, w1, b1
 };
